requirements.txt
PyJWT==2.7.0
pydantic==1.10.11
fastapi==0.99.1
SQLAlchemy==2.0.18
uvicorn==0.22.0
asyncpg==0.28.0